+++ tests/Unit/Form/FormExtensionTypesTest.php
@@ -8,6 +8,7 @@ use CraftCms\Cms\Form\Controls\Address;
 use CraftCms\Cms\Form\Controls\Choice;
 use CraftCms\Cms\Form\Controls\Color;
 use CraftCms\Cms\Form\Controls\Date;
+use CraftCms\Cms\Form\Controls\ElementSelect;
 use CraftCms\Cms\Form\Controls\IconPicker;
 use CraftCms\Cms\Form\Controls\Lightswitch;
 use CraftCms\Cms\Form\Controls\Link;
@@ -45,6 +46,7 @@ it('registers core and plugin Node and Control types separately', function () {
             Choice::class,
             Color::class,
             Date::class,
+            ElementSelect::class,
             IconPicker::class,
             Lightswitch::class,
             Link::class,